Closes #4: Use multi-valued RDN in DN to avoid account hijacking.
This adds support for sanitizing multi-valued RDNs and changes the
default construction of a user's DN to include such a multi-valued RDN.
In so doing, we now embed the WordPress user's auto-incrementing numeric
user ID value as part of their DN, which prevents the possibility of a
given account being re-used by multiple different registrations.

Imagine the following scenario:

1. A user registers as "UserA" and is given an ID number of 2.
2. Sometime in the future, a Super Admin deletes this user from WP.
3. Afterwards, a distinct human registers an account called "UserA"

The second human to register the same username ("UserA") is now
potentially in control over any SSO-linked accounts using the given
username, resulting in a possible account hijack.

This is possible only if a Super Admin manually removes the WordPress
user record from WordPress's database, but this isn't that far-fetched
of a possibility. Embedding the auto-incrementing user ID into the LDAP
DN means that when (3) happens in the timeline above, the entity
associated with this second human's registration will be distinct from
the first human's registration, despite both users having the same `uid`
value in the LDAP DIT (i.e., `user_login` on the WordPress side).

This change drops support for Microsoft Active Directory, as MS AD does
not support multi-valued RDNs. That being said, fuck Microsoft anyway.